feat: Hero section change
